(WIP) Add the ability to add a remote NFS file system at run time as a local file system

+++ cm/util/misc.py
@@ -568,8 +568,8 @@ def get_file_from_bucket(conn, bucket_name, remote_filename, local_file, validat
         except S3ResponseError as e:
             log.debug("Failed to get file '%s' from bucket '%s': %s" % (
                 remote_filename, bucket_name, e))
-            os.remove(local_file)
-                      # Don't leave a partially downloaded or touched file
+            if os.path.exists(local_file):
+                os.remove(local_file)  # Don't leave a partially downloaded or touched file
             return False
     else:
         log.debug("Bucket '%s' does not exist, did not get remote file '%s'" % (
